requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
faiss-cpu==1.7.4
sentence-transformers==2.2.2
httpx==0.25.2
python-dotenv==1.0.0
aiosqlite==0.19.0
